Solved AoC2020 Day 6 Part 2

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day06/Day06Test.kt
@@ -15,7 +15,7 @@ class Day06Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         CustomCustoms(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
@@ -29,7 +29,7 @@ class Day06Test {
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day06/day06part2sample.txt", 6L)
+                Arguments.of("de/havox_design/aoc2020/day06/day06part2sample.txt", 6)
             )
     }
 }
